tighten up quoting of shell variables overall
I grepped for all unquoted variables that I could find and fixed them.

Most of these are not going to cause security problems or other bugs, but
the rule of thumb re quoting shell variables is that if you don't want to
have to worry about it, always quote it. Especially when your program
has millions of users and is starting to be used in environments it was
never really expected to be used in.

(Left unquoted: Variables used in "set -- $var" and "for $var in", and
a very few other special cases.)

r62230