Avoid false positives in MS-DOS detection by also looking for autoexec.bat. Closes: #663540
/dos is not an unheard of directory to have on a linux system -- we used
to put our dual-boot mount point there in 1995 -- and when mounted by
grub-mount, the filesystem type is not available to limit detection of
MS-DOS to only relevant filesystems.

Work around this problem by also requiring an /autoexec.bat or a
/config.sys. While not all MS-DOS systems have one of these files,
any MS-DOS system that a user is likely to care about showing up
in their boot menu probably does.
(IIRC not all MS-DOS systems even have a /dos!)

I also took a brief look at the other tests that allows fuse mount through.
Many of them, like 83haiku check for enough different items that the
chances of a false positive are very low even absent a filesystem type
confirmation. The most worrying are 20macosx, which only checks for
/mach_kernel, and 30utility, which only checks for /f11.sys.

Those should probably be tightened down, or a way found to check the actual
filesystem type when using grub-mount.